refactor(lib): Move ZettelFileName from the store hierarchy to the entities hierarchy
